Removed AddQuotedParameter and AddUnquotedParameter from ICommandBuilder and replaced them with AddParameter Changed CommandBuilder implementation to match the interface Defined ILinkRepository and implemented SQLiteLinkRepository Added tests for SQLiteLinkRepository read methods

+++ Gnosis/src/Gnosis.Core/ILinkType.cs
@@ -7,7 +7,7 @@ namespace Gnosis.Core
 {
     public interface ILinkType
     {
-        long Id { get; }
+        int Id { get; }
         string Name { get; }
     }
 }
